Moved performance keyword unit tests to a separate file. Also, fixed the broken tests caused by the refactor.
