+++ cumulusci/robotframework/tests/test_salesforce.py
@@ -1,7 +1,6 @@
 from unittest import mock
 
 import pytest
-import responses
 from SeleniumLibrary.errors import ElementNotFound
 
 from cumulusci.robotframework.Salesforce import Salesforce
@@ -78,35 +77,3 @@ class TestKeyword_breakpoint:
     def test_breakpoint(self, mock_robot_context):
         """Verify that the keyword doesn't raise an exception"""
         assert self.sflib.breakpoint() is None
-
-
-class TestKeyword_elapsed_time_for_last_record:
-    @responses.activate
-    def test_elapsed_time_for_last_record__query_empty(self):
-        sflib = Salesforce()
-        records = {"records": []}
-
-        with mock.patch.object(Salesforce, "cumulusci") as cumulusci:
-            cumulusci.sf.query_all.return_value = records
-            with pytest.raises(Exception) as e:
-                sflib.elapsed_time_for_last_record("FOO", "Bar", "Baz", "Baz")
-            assert "Matching record not found" in str(e.value)
-
-    @responses.activate
-    def test_elapsed_time_for_last_record__query_returns_result(self):
-        sflib = Salesforce()
-        records = {
-            "records": [
-                {
-                    "CreatedDate": "2020-12-29T10:00:01.000+0000",
-                    "CompletedDate": "2020-12-29T10:00:04.000+0000",
-                }
-            ],
-        }
-
-        with mock.patch.object(Salesforce, "cumulusci") as cumulusci:
-            cumulusci.sf.query_all.return_value = records
-            elapsed = sflib.elapsed_time_for_last_record(
-                "AsyncApexJob", "CreatedDate", "CompletedDate", "CompletedDate"
-            )
-            assert elapsed == 3
